Keeping elevation value in extract and prepare
This commit introduces changes needed for returning the elevation
information associated with each node in the ViaRoute plugin.

The elevation is read from OSM as a tag and is stored as a new field in
the ExternalMemoryNode structure. It is stored to OSRM at the end of the
extract step and kept throuh the prepare step. When the
InternalDataFacade data structure is built, the elevation information
for all the nodes is stored in a separate list which can be queried with
the new method: GetElevationOfNode(...)