Elevation
- Import elevation from osm node tag 'ele';
- Optional elevation in JSON and GPX formats;
- Phantom nodes elevation set early;
- Elevation tests with cucumber -p elevation.

The elevation is stored as an int.
Using a smaller structure would lead to misalignement or no gain.

The routing algorithm is unchanged.

The elevation feature must be enabled at compile time with:
cmake option -DOSRM_HAS_ELEVATION:BOOL=ON and each tool must be passed
the '-e 1' parameter to activate it. Finally, the viaroute request may
either include the 'elevation=true' parameter to get elevation, set it
to false or skip it to get default behaviour.

The elevation is read as meters from the osm data.
Since we must store it as an int, we get a 1cm precision.
It is then exported as 10^4 meters in JSON.