Provide a Makefile wrapping the beast that CMake is
This provides an easy to use Makefile, that wraps the build process
behind a simple `make` command. In particular, it builds in release mode
and in parallel by default. It also separates release builds from debug
builds, in the build directory.

Discussion: should be update all our documentation, readmes and wikis,
to simply invoke `make`?

To be done:

- the raster unit tests hardcode "../unit_tests/" as path prefix; this
  breaks with a build directory such as "build/Release/"

- the cucumber setup assumes the osrm binaries are located in "build/"