Reproducible development environment via nix-shell
This adds a basix `shell.nix` file (similar to requirements.txt for
Python's pip), for reliable and reproducible builds.

Enter the environment:

    nix-shell

Or enter the environment without system binaries available:

    nix-shell --pure

This is similar to a negation of `pyvenv --system-side-packages`.

On first invocation all dependencies (i.e. compiler, build system,
dependencies) are downloaded by means of resolving a dependency tree
down to libc. Because this tree can be hashed at nodes, binary packages
can be fetched for the common packages and verified against hashes.

In case we have ABI mismatches, e.g. because we want to use gcc5 with
its stdlib, we can override the dependencies' environment. Nix then looks
for binary caches for those packages with the gcc5 stdlib and if they
are not available, build the dependency tree with the new environment.

    buildInputs = map (pkg: pkg.override { stdenv = overrideCC stdenv gcc5; }) [ gcc5 cmake .. ];

Note: this is a first try in providing a development environment that
allows for reproducible builds. Building osrm-backend works, but I did
not (yet?) add all dependencies e.g. for the tools (gdal etc.).

Please see the quickstart guide below, and give it a try.

Resources:

- http://nixos.org/nix/
- https://nixos.org/nix/manual/#chap-quick-start
- https://nixos.org/nixpkgs/manual/