Create pidfile once osrm-routed is ready to serve request.
Given that osrm-routed can take some time to load its data and be ready
to serve request create a pidfile only once requests will be served.
This is useful when hot reloading osrm-routed: a supervising process (eg
rainbow-saddle) can start a new osrm-routed process and watch for the
creation of the new pid file. Then the old osrm-routed can be shutdown.