Prevents typedefs.hpp from polluting the user's namespace, resolves #2940
`typefefs.hpp` defined some types and type-aliases (such as node / way
id for example) but did so without a namespace, polluting the global
one. This can cause name clashes when using / embedding libosrm.

This changeset namespaces types and type aliases in `typedefs.hpp` into
the `osrm` namespace and adapts some callsites.

Because the `libosrm` headers we ship to clients depend on
`typedefs.hpp` and clients could already use it this change is breaking
the `libosrm` v5 API.